finished the page Aziende OP completed (3,4)

+++ app/src/main/java/concessionario/db/tables/Azienda.java
@@ -1,19 +1,19 @@
 package concessionario.db.tables;
 
 public class Azienda {
-    private final int partita_iva;
+    private final long partita_iva;
     private final String nome;
     private final String sede;
     private final int fatturato;
     
-    public Azienda(int partita_iva, String nome, String sede, int fatturato) {
+    public Azienda(long partita_iva, String nome, String sede, int fatturato) {
         this.partita_iva = partita_iva;
         this.nome = nome;
         this.sede = sede;
         this.fatturato = fatturato;
     }
 
-    public int getPartita_iva() {
+    public long getPartita_iva() {
         return partita_iva;
     }
